+++ server/common/getWorkLocation.js
@@ -1,4 +1,4 @@
-const capitalizeLocation = require('./capitalizeLocation');
+const fixLocationName = require('./fixLocationName');
 
 function getWorkLocation(propertyLocation, locationList, hidePostAddress = true) {
     if (propertyLocation) {
@@ -12,17 +12,17 @@ function getWorkLocation(propertyLocation, locationList, hidePostAddress = true)
     const workLocations = [];
     for (let i = 0; i < locationList.length; i += 1) {
         if (locationList[i].city && hidePostAddress) {
-            workLocations.push(capitalizeLocation(locationList[i].city));
+            workLocations.push(fixLocationName(locationList[i].city));
         } else if (locationList[i].postalCode) {
             let tmp = locationList[i].address ? `${locationList[i].address}, ` : '';
-            tmp += `${locationList[i].postalCode} ${capitalizeLocation(locationList[i].city)}`;
+            tmp += `${locationList[i].postalCode} ${fixLocationName(locationList[i].city)}`;
             workLocations.push(tmp);
         } else if (locationList[i].municipal) {
-            workLocations.push(capitalizeLocation(locationList[i].municipal));
+            workLocations.push(fixLocationName(locationList[i].municipal));
         } else if (locationList[i].county) {
-            workLocations.push(capitalizeLocation(locationList[i].county));
+            workLocations.push(fixLocationName(locationList[i].county));
         } else if (locationList[i].country) {
-            workLocations.push(capitalizeLocation(locationList[i].country));
+            workLocations.push(fixLocationName(locationList[i].country));
         }
     }